+++ src/components/Brand/index.ts
@@ -1,0 +1,2 @@
+export { default as Brand } from './Brand';
+export type { default as BrandProps } from './Brand';